# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/{hotelId}/rooms": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "description": "Retrieve all rooms available for a specific hotel.",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the hotel"
          }
        ],
        "responses": {
          "200": {
            "description": "Rooms retrieved successfully.",
            "content": {
              "text/html": {}
            }
          },
          "404": {
            "description": "Hotel not found or has no rooms.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "Hotel Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/hotels": {
      "post": {
        "tags": [
          "Hotels"
        ],
        "description": "Creates a new hotel with rooms. Admin access only.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Hotel details including name, location, and rooms",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Hotel"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Hotel and rooms created successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Missing required fields."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/{hotelId}/reservations": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "description": "Retrieve all reservations for a specific hotel.",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Hotel ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Reservations retrieved successfully.",
            "content": {
              "text/html": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/hotels/{hotelId}/rate": {
      "post": {
        "tags": [
          "Hotels"
        ],
        "description": "Allows a user to rate a hotel.",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the hotel"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Rating from 1 to 5",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "rating": {
                  "type": "number",
                  "example": 5
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel rated successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Invalid rating."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Hotel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Example Hotel"
        },
        "location": {
          "type": "string",
          "example": "Example City"
        },
        "rooms": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "capacity": {
                "type": "number",
                "example": 2
              },
              "price": {
                "type": "number",
                "example": 100
              }
            }
          }
        }
      },
      "required": [
        "name",
        "location",
        "rooms"
      ]
    },
...
  }
}